feat(x/auth/tx): extend app module + docs
