+++ x/auth/tx/config/depinject.go
@@ -15,8 +15,6 @@ import (
 	txconfigv1 "cosmossdk.io/api/cosmos/tx/config/v1"
 	"cosmossdk.io/core/address"
 	"cosmossdk.io/core/appmodule"
-	appmodulev2 "cosmossdk.io/core/appmodule/v2"
-	"cosmossdk.io/core/transaction"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/depinject/appconfig"
 	"cosmossdk.io/x/auth/ante"
@@ -58,15 +56,16 @@ type ModuleInputs struct {
 	AccountAbstractionKeeper ante.AccountAbstractionKeeper      `optional:"true"`
 	CustomSignModeHandlers   func() []txsigning.SignModeHandler `optional:"true"`
 	CustomGetSigners         []txsigning.CustomGetSigner        `optional:"true"`
+	ExtraTxValidators        []TxValidator                      `optional:"true"`
 }
 
 type ModuleOutputs struct {
 	depinject.Out
 
+	Module          appmodule.AppModule   // This is only useful for chains using server/v2. It setup tx validators that don't belong to other modules.
+	BaseAppOption   runtime.BaseAppOption // This is only useful for chains using baseapp. Server/v2 chains use TxValidator.
 	TxConfig        client.TxConfig
 	TxConfigOptions tx.ConfigOptions
-	BaseAppOption   runtime.BaseAppOption // This is only useful for chains using baseapp. Server/v2 chains use TxValidator.
-	Module          appmodule.AppModule
 }
 
 func ProvideProtoRegistry() txsigning.ProtoFileResolver {
@@ -150,9 +149,13 @@ func ProvideModule(in ModuleInputs) ModuleOutputs {
 		ante.DefaultSigVerificationGasConsumer,
 		in.AccountAbstractionKeeper,
 	)
-	appModule := AppModule{svd}
 
-	return ModuleOutputs{TxConfig: txConfig, TxConfigOptions: txConfigOptions, BaseAppOption: baseAppOption, Module: appModule}
+	return ModuleOutputs{
+		Module:          NewAppModule(svd, in.ExtraTxValidators...),
+		TxConfig:        txConfig,
+		TxConfigOptions: txConfigOptions,
+		BaseAppOption:   baseAppOption,
+	}
 }
 
 func newAnteHandler(txConfig client.TxConfig, in ModuleInputs) (sdk.AnteHandler, error) {
@@ -232,23 +235,3 @@ func metadataExists(err error) error {
 
 	return err
 }
-
-var (
-	_ appmodulev2.AppModule                      = AppModule{}
-	_ appmodulev2.HasTxValidator[transaction.Tx] = AppModule{}
-)
-
-type AppModule struct {
-	sigVerification ante.SigVerificationDecorator
-}
-
-// TxValidator implements appmodule.HasTxValidator.
-func (a AppModule) TxValidator(ctx context.Context, tx transaction.Tx) error {
-	return a.sigVerification.ValidateTx(ctx, tx)
-}
-
-// IsAppModule implements appmodule.AppModule.
-func (a AppModule) IsAppModule() {}
-
-// IsOnePerModuleType implements appmodule.AppModule.
-func (a AppModule) IsOnePerModuleType() {}
